Fix backup

+++ rogu/resources.py
@@ -505,8 +505,10 @@ class File(_UgorResource):
 
         if not self.path.parent.exists():
             self.path.parent.mkdir(parents=True)
-        backup = self.path.parent / f'.{self.path.name}~'
-        shutil.copy(self.path, backup)
+        if self.path.exists():
+            backup = self.path.parent / f'.{self.path.name}~'
+            shutil.copy(self.path, backup)
+            debug(f'Backed up to {backup.name}')
         self.path.write_bytes(content)
 
     def upload(self, force: bool = False):
